updated schema file for database slug issues.
As we had some database issues which made application stopped after running for 5 minutes and when we wanted to publish docs on server. I resolved that issue. But still have some bugs. Because of that bug controller is not fetching data from database and so I am working on it.